# docs/forensics/openapi.yaml
openapi: 3.1.0
info:
  title: BankSystem Financial Forensics API
  version: 1.0.0
  description: Admin API for deterministic investigation, verification, case management, replay and evidence-grounded Copilot.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /admin/forensics/verification/check/{transactionId}:
    post:
      tags: [Verification]
      summary: Run deterministic invariant verification
      parameters:
        - $ref: '#/components/parameters/TransactionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses: { '200': { $ref: '#/components/responses/Success' }, '409': { $ref: '#/components/responses/Error' } }
  /admin/forensics/verification/runs/{runId}:
    get:
      tags: [Verification]
      summary: Get verification run and rule results
      parameters: [ { $ref: '#/components/parameters/RunId' } ]
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
  /admin/forensics/twin/forks:
    post:
      tags: [Replay]
      summary: Create an isolated sanitized fork with TTL and quota
      responses: { '200': { $ref: '#/components/responses/Success' }, '429': { $ref: '#/components/responses/Error' } }
  /admin/forensics/twin/replays:
    post:
      tags: [Replay]
      summary: Run a confirmed allowlisted scenario against the deployed commit
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      responses: { '200': { $ref: '#/components/responses/Success' }, '409': { $ref: '#/components/responses/Error' } }
  /admin/forensics/twin/runs/{runId}:
    get:
      tags: [Replay]
      summary: Return replay lifecycle and before/after verification metadata
      parameters: [ { $ref: '#/components/parameters/RunId' } ]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    TransactionId: { name: transactionId, in: path, required: true, schema: { type: string, format: uuid } }
    RunId: { name: runId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 120 } }
  responses:
    Success:
      description: ApiResponse success envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    Error:
      description: Stable business error envelope without stack trace or sensitive detail
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
  schemas:
    ApiResponse:
      type: object
      required: [success, timestamp]
      properties:
        success: { type: boolean }
        data: {}
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
        timestamp: { type: string, format: date-time }